Add files via upload
sc dropmedia allows people to add a folder or individual files into a menu to allow for easy switching between different types of input media. Sc loadpatch allows people to add folder of max patches into a menu which can then be used to load in each patch into a bpatcher object. These are just prototypes of ideas that would be useful for media composition